Move setup methods back into interface classes for MFEM LOR solvers that do not use the MFEM LOR solver template
