+++ framework/include/mfem/solvers/MFEMHyprePCG.h
@@ -28,7 +28,7 @@ public:
 
   void Update() override
   {
-    SetupLOR(*this, *_equation_system);
+    SetLORSolver(*this, *_equation_system);
     // Reset preconditioner if it has had an LOR update
     if (IsLOR(*this) && !_lor)
       SetPreconditioner(static_cast<mfem::HyprePCG &>(GetSolver()));
