+++ framework/src/mfem/solvers/MFEMHypreAMS.C
@@ -64,7 +64,7 @@ MFEMHypreAMS::Update()
     if (_mfem_fespace.getFESpace()->GetMesh()->GetElement(0)->GetGeometryType() !=
         mfem::Geometry::Type::CUBE)
       mooseError("LOR HypreAMS Solver only supports hex meshes.");
-    Moose::MFEM::LORLinearSolverBase<mfem::HypreAMS>::SetupLOR(*this, *_equation_system);
+    SetLORSolver(*this, *_equation_system);
   }
 }
