+++ framework/src/mfem/solvers/MFEMLORLinearSolverBase.C
@@ -20,13 +20,13 @@ InputParameters
 LORLinearSolverBase<MFEMSolverType>::validParams()
 {
   InputParameters params = LinearSolverBase::validParams();
-  params.addParam<bool>("low_order_refined", false, "Set usage of Low-Order Refined solver.");
+  params += LORInterface::validParams();
   return params;
 }
 
 template <class MFEMSolverType>
 LORLinearSolverBase<MFEMSolverType>::LORLinearSolverBase(const InputParameters & parameters)
-  : Moose::MFEM::LinearSolverBase(parameters), _lor{parameters.get<bool>("low_order_refined")}
+  : LinearSolverBase(parameters), LORInterface(parameters)
 {
 }
 
@@ -40,77 +40,14 @@ LORLinearSolverBase<MFEMSolverType>::Update()
     if (_lor && GetPreconditioner())
       mooseError("LOR solver cannot take a preconditioner");
     if (_lor)
-      LORLinearSolverBase<MFEMSolverType>::SetupLOR(*this, *_equation_system);
+      LORLinearSolverBase<MFEMSolverType>::SetLORSolver(*this, *_equation_system);
   }  
 }
 
 template <class MFEMSolverType>
-bool
-LORLinearSolverBase<MFEMSolverType>::IsLOR(LinearSolverBase & solver) const
-{
-  LORLinearSolverBase<MFEMSolverType> * lor_preconditioner = GetPreconditionerLORInterface(solver);
-  return _lor || (lor_preconditioner && lor_preconditioner->IsLOR(*solver.GetPreconditioner()));
-}
-
-template <class MFEMSolverType>
-LORLinearSolverBase<MFEMSolverType> *
-LORLinearSolverBase<MFEMSolverType>::GetPreconditionerLORInterface(LinearSolverBase & solver) const
-{
-  return dynamic_cast<LORLinearSolverBase<MFEMSolverType> *>(solver.GetPreconditioner());
-}
-
-template <class MFEMSolverType>
 void
-LORLinearSolverBase<MFEMSolverType>::CheckSpectralEquivalence(mfem::ParBilinearForm & blf) const
-{
-  if (auto fec = dynamic_cast<const mfem::H1_FECollection *>(blf.FESpace()->FEColl()))
-  {
-    if (fec->GetBasisType() != mfem::BasisType::GaussLobatto)
-      mooseError("Low-Order-Refined solver requires the FESpace basis to be GaussLobatto "
-                 "for H1 elements.");
-  }
-  else if (auto fec = dynamic_cast<const mfem::ND_FECollection *>(blf.FESpace()->FEColl()))
-  {
-    if (fec->GetClosedBasisType() != mfem::BasisType::GaussLobatto ||
-        fec->GetOpenBasisType() != mfem::BasisType::IntegratedGLL)
-      mooseError("Low-Order-Refined solver requires the FESpace closed-basis to be GaussLobatto "
-                 "and the open-basis to be IntegratedGLL for ND elements.");
-  }
-  else if (auto fec = dynamic_cast<const mfem::RT_FECollection *>(blf.FESpace()->FEColl()))
-  {
-    if (fec->GetClosedBasisType() != mfem::BasisType::GaussLobatto ||
-        fec->GetOpenBasisType() != mfem::BasisType::IntegratedGLL)
-      mooseError("Low-Order-Refined solver requires the FESpace closed-basis to be GaussLobatto "
-                 "and the open-basis to be IntegratedGLL for RT elements.");
-  }
-}
-
-template <class MFEMSolverType>
-void
-LORLinearSolverBase<MFEMSolverType>::SetupLOR(Moose::MFEM::EquationSystem & equation_system)
-{
-  if (equation_system.isComplex())
-    mooseError("LOR solve is not supported for complex equation systems.");
-  if (equation_system.GetTestVarNames().size() > 1)
-    mooseError("LOR solve is only supported for single-variable systems");
-
-  const auto & test_var_name = equation_system.GetTestVarNames().at(0);
-  const auto & trial_var_name = equation_system.GetTrialVarNames().at(0);
-  mfem::ParGridFunction & trial_gf = equation_system.getGridFunction(trial_var_name);
-  _a = &equation_system.GetBilinearForm(test_var_name);
-  CheckSpectralEquivalence(*_a);
-
-  _ess_bdr_markers.SetSize(trial_gf.ParFESpace()->GetParMesh()->bdr_attributes.Max());
-  _ess_bdr_markers = 0;
-  equation_system.ApplyEssentialBC(trial_var_name, trial_gf, _ess_bdr_markers);
-
-  _a->ParFESpace()->GetEssentialTrueDofs(_ess_bdr_markers, _ess_tdofs);
-}
-
-template <class MFEMSolverType>
-void
-LORLinearSolverBase<MFEMSolverType>::SetupLOR(LinearSolverBase & solver_base,
-                                              Moose::MFEM::EquationSystem & equation_system)
+LORLinearSolverBase<MFEMSolverType>::SetLORSolver(LinearSolverBase & solver_base,
+                                                  Moose::MFEM::EquationSystem & equation_system)
 {
   if (_lor)
   {
@@ -123,8 +60,8 @@ LORLinearSolverBase<MFEMSolverType>::SetupLOR(LinearSolverBase & solver_base,
 
 template <>
 void
-LORLinearSolverBase<mfem::HypreGMRES>::SetupLOR(LinearSolverBase & solver_base,
-                                          Moose::MFEM::EquationSystem & equation_system)
+LORLinearSolverBase<mfem::HypreGMRES>::SetLORSolver(LinearSolverBase & solver_base,
+                                                    Moose::MFEM::EquationSystem & equation_system)
 {
   if (_lor)
   {
@@ -138,8 +75,8 @@ LORLinearSolverBase<mfem::HypreGMRES>::SetupLOR(LinearSolverBase & solver_base,
 
 template <>
 void
-LORLinearSolverBase<mfem::HypreFGMRES>::SetupLOR(LinearSolverBase & solver_base,
-                                          Moose::MFEM::EquationSystem & equation_system)
+LORLinearSolverBase<mfem::HypreFGMRES>::SetLORSolver(LinearSolverBase & solver_base,
+                                                     Moose::MFEM::EquationSystem & equation_system)
 {
   if (_lor)
   {
@@ -153,8 +90,8 @@ LORLinearSolverBase<mfem::HypreFGMRES>::SetupLOR(LinearSolverBase & solver_base,
 
 template <>
 void
-LORLinearSolverBase<mfem::HyprePCG>::SetupLOR(LinearSolverBase & solver_base,
-                                       Moose::MFEM::EquationSystem & equation_system)
+LORLinearSolverBase<mfem::HyprePCG>::SetLORSolver(LinearSolverBase & solver_base,
+                                                  Moose::MFEM::EquationSystem & equation_system)
 {
   if (_lor)
   {
@@ -164,17 +101,6 @@ LORLinearSolverBase<mfem::HyprePCG>::SetupLOR(LinearSolverBase & solver_base,
     SetSolverParameters(lor_solver->GetSolver());
     solver_base.SetSolver(lor_solver);
   }
-}
-
-template <>
-void
-LORLinearSolverBase<Moose::MFEM::MatrixFreeAMS>::SetupLOR(LinearSolverBase & solver_base,
-                                                          Moose::MFEM::EquationSystem & equation_system)
-{
-  SetupLOR(equation_system);
-  auto & matrix_free_ams = static_cast<Moose::MFEM::MatrixFreeAMS &>(*_solver);
-  matrix_free_ams.SetBilinearForm(*_a);
-  matrix_free_ams.SetBoundaryMarkers(_ess_bdr_markers);
 }
 
 } // namespace Moose::MFEM
